Removing the final compiler warning when building on Win75 -- thank you, Antonio Mattos.
git-svn-id: http://svn.code.sf.net/p/freeglut/code/trunk/freeglut/freeglut@731 7f0cb862-5218-0410-a997-914c9d46530a

+++ freeglut/src/freeglut_main.c
@@ -2088,7 +2088,7 @@ LRESULT CALLBACK fgWindowProc( HWND hWnd, UINT uMsg, WPARAM wParam,
 
             GetKeyboardState( state );
 
-            if( ToAscii( wParam, 0, state, code, 0 ) == 1 )
+            if( ToAscii( (UINT)wParam, 0, state, code, 0 ) == 1 )
                 wParam=code[ 0 ];
 
             INVOKE_WCB( *window, KeyboardUp,
